+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-## v0.0.4 (UNRELEASED)
-* Allow for listing of current config and context using `-l` and `-lc` flags.
+## v0.0.4 (September 25, 2023)
+* Allow for listing of current config and context.
 
 ## v0.0.3 (September 23, 2023)
 * Added support for default config at `~/.kube/config`.
-* Added context switcher using `-c` flag.
+* Added context switcher.
 
 ## v0.0.2 (September 22, 2023)
 * Added `KXD_MATCHER` environment variable for specifying different config matchers.